requirements.txt
asgiref==3.10.0
Django==5.2.4
django-jazzmin==3.0.1
django-markdownx==4.0.9
django-widget-tweaks==1.5.0
Markdown==3.8.2
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
django-allauth==65.10.0
bleach==6.2.0
webencodings==0.5.1
